+++ backend/wrangler.jsonc
@@ -5,8 +5,8 @@
   "compatibility_date": "2026-03-17",
   "vars": {
     "GOOGLE_CLIENT_ID": "457796621894-7krpa8n09mekpj2cdd952jjkf1c3ookt.apps.googleusercontent.com",
-    "FRONTEND_URL": "fantasywiki.pages.dev"
-  }
+    "FRONTEND_URL": "fantasywiki.pages.dev",
+  },
   // "compatibility_flags": [
   //   "nodejs_compat"
   // ],
