+++ frontend/src/components/AppLogo.vue
