+++ frontend/src/components/ArticleDetail.vue
