+++ frontend/src/components/homePage/AppStats.vue
