+++ frontend/src/components/homePage/LeaderboardPreview.vue
@@ -203,7 +203,7 @@ const getRankClass = (rank: number) => {
 };
 </script>
 
-<style scoped src="./home-page.css"></style>
+<style scoped src="home-page.css"></style>
 <style scoped>
 ion-chip {
   font-weight: 500;
