+++ frontend/src/tests/homePage/AppFeature.spec.ts
@@ -1,14 +1,13 @@
 import { describe, it, expect } from "vitest";
 import { mount } from "@vue/test-utils";
 import router from "@/router/index";
-import CTASection from "@/views/home-page/CTASection.vue";
+import AppFeature from "@/components/homePage/AppFeature.vue";
 
-describe("home-page/CTASection.vue", () => {
+describe("AppFeature.vue", () => {
   it("should mount without any console errors or warnings", async () => {
     router.push("/");
     await router.isReady();
-
-    const wrapper = mount(CTASection, {
+    const wrapper = mount(AppFeature, {
       global: {
         plugins: [router],
       },
