+++ frontend/src/tests/homePage/CTASection.spec.ts
@@ -1,13 +1,14 @@
 import { describe, it, expect } from "vitest";
 import { mount } from "@vue/test-utils";
 import router from "@/router/index";
-import HowItWorks from "@/views/home-page/HowItWorks.vue";
+import CTASection from "@/components/homePage/CTASection.vue";
 
-describe("HowItWorks.vue", () => {
+describe("home-page/CTASection.vue", () => {
   it("should mount without any console errors or warnings", async () => {
     router.push("/");
     await router.isReady();
-    const wrapper = mount(HowItWorks, {
+
+    const wrapper = mount(CTASection, {
       global: {
         plugins: [router],
       },
