+++ frontend/src/tests/homePage/LeaderboardPreview.spec.ts
@@ -1,13 +1,13 @@
 import { describe, it, expect } from "vitest";
 import { mount } from "@vue/test-utils";
 import router from "@/router/index";
-import ArticleLeaderboard from "@/views/home-page/ArticleLeaderboard.vue";
+import LeaderboardPreview from "@/components/homePage/LeaderboardPreview.vue";
 
-describe("ArticleLeaderboard.vue", () => {
+describe("home-page/LeaderboardPreview.vue", () => {
   it("should mount without any console errors or warnings", async () => {
     router.push("/");
     await router.isReady();
-    const wrapper = mount(ArticleLeaderboard, {
+    const wrapper = mount(LeaderboardPreview, {
       global: {
         plugins: [router],
       },
